Swap position of mouse actions and project actions
