+++ libs/ui-toolbar/src/main/kotlin/mono/html/toolbar/view/nav/Components.kt
@@ -7,14 +7,25 @@
 package mono.html.toolbar.view.nav
 
 import androidx.compose.runtime.Composable
+import mono.ui.compose.ext.classes
 import org.jetbrains.compose.web.dom.ContentBuilder
 import org.jetbrains.compose.web.dom.Div
 import org.w3c.dom.HTMLDivElement
 
 @Composable
-internal fun ToolbarContainer(content: ContentBuilder<HTMLDivElement>? = null) {
+internal fun ToolbarContainer(
+    addSpaceLeft: Boolean = false,
+    addSpaceRight: Boolean = false,
+    content: ContentBuilder<HTMLDivElement>? = null
+) {
     Div(
-        attrs = { classes("toolbar-container") }
+        attrs = {
+            classes(
+                "toolbar-container" to true,
+                "add-space-left" to addSpaceLeft,
+                "add-space-right" to addSpaceRight
+            )
+        }
     ) {
         content?.invoke(this)
     }
