+++ src/main/sass/style/header/action-toolbar.scss
@@ -4,6 +4,14 @@
     height: 100%;
     align-items: center;
 
+    &.add-space-left {
+        margin-left: 4px;
+    }
+
+    &.add-space-right {
+        margin-right: 4px;
+    }
+
     .app-icon-container {
         display: flex;
         align-items: center;
