+++ src/main/sass/style/header/style.scss
@@ -1,4 +1,4 @@
 @import 'header';
 @import 'working-file';
 @import 'mouse-actions';
-@import 'right-toolbar';
+@import 'action-toolbar';
